Do not update HotspotContainer rect when scene layer effects change.
Since scenes may now have an arbitrary number of layers (even zero),
syncing the HotspotContainer's rect with one of the layers seems
awkward.

We may add this back in the future, when we properly support setting the
rect for all layers in a scene at once. For now, it's simpler to assume
a full rect.