Correctly enable/disable attributes when switching renderers.
Enabled attributes are a global setting. Previously, shader attributes
were enabled when creating the shader, and never disabled them. Hence,
when switching to a shader with a smaller number of attributes, some
attributes remained enabled even though they weren't bound to anything.
This caused flickering when switching from a cube into an equirect.

The fix is to have each renderer enable the attributes it needs in
startLayer and disable them in endLayer.

Closes #171.