Make WebGL rendering resilient to device pixel ratio changes.
The device pixel ratio is currently necessary when updating the stage size
and the viewport. However, since changes to the device pixel ratio are not
necessarily accompanied by a window resize event (for example, when moving
the window across screens), the viewport may use a different value than the
stage size, which breaks the rendering.

The solution is to read the device pixel ratio only when the stage size
changes, and leverage the `drawingBufferWidth` and `drawingBufferHeight`
properties of the WebGLRenderingContext at other times. On Flash and CSS,
the `clientWidth` and `clientHeight` properties of the stage element can be
used instead.

Because a changing device pixel ratio makes absolute viewport sizes
ambiguous, it is also a bad idea for the rendering pipeline to pass around
non-normalized rects (and indeed, the documentation already suggested that
Renderer#startLayer and Renderer#endLayer receive normalized rects.
Therefore, calcRect() is changed to always return a normalized rect. We also
take the opportunity to clean up the returned structure (which contains a
lot of redundant information) and document it properly.

Fixes #196.