Remove the `stage` argument from the Layer constructor.
Storing a reference to a stage in a layer is problematic, because it is
not kept in sync: the user must still add the layer into the stage, and
there is no way to update the reference if the layer is removed from the
stage or added into a different stage. Therefore, calling Layer#stage()
offers no guarantee that the return value is correct.

Since a layer is an otherwise immutable data structure, it seems sensible
to remove the stage reference entirely. Unfortunately, this is a breaking
change for clients that instantiate Layer directly.