Create renderer eagerly when layer is added.
This moves renderer creation out of the render path. A layer's underlying
geometry and view types cannot change after creation, so there's no need to
re-check that the renderer type is correct every time we render.

Also switch from using a Map to store renderers (which is awkward) in favor
of an array, which should be efficient enough as the number of distinct
renderers is very small.