Implement progressive loading.
Previously, the stage always attempted to load and render the optimal
resolution level for the current view. When zooming through multiple
resolution levels in a gigapixel panorama, this results in a poor visual
effect, as a very low resolution image must stand in for textures that
are yet to load.

A more pleasing experience can be obtained by loading every level (for
the area of interest) in increasing resolution order. This comes at the
cost of increased bandwidth consumption, but seems to be the right
tradeoff for gigapixel panoramas. For regular panoramas, the number of
levels is small, so it shouldn't have a big impact.

Fixes #99.