Refactor the collection classes.
Abandon the requirement of passing in custom equals() and hash() functions
in favor of prototype methods; this removes a bit of boilerplate from the
tile classes, which previously exposed both instance and static methods.
Also rename `each` as `forEach` in line with language conventions.

Also fix a subtle bug in LruSet and LruMap, whereby removing an element
before the data structure has grown to full size would cause a persistent
hole.